requirements.txt
# Install package + optional extras: pip install -e ".[api,ui,dev]"
# Or minimal runtime:
numpy>=1.26,<3
pandas>=2.1,<3
matplotlib>=3.8,<4
torch>=2.2,<3
scikit-learn>=1.4,<2
PyYAML>=6.0.1,<7
joblib>=1.3,<2
